Improved standards.

+++ CrunchyPeanutButter.Core/Commands/Bars/DeleteBar/DeleteBarCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using CrunchyPeanutButter.Core.Abstractions;
 using MediatR;
@@ -7,9 +8,9 @@ namespace CrunchyPeanutButter.Core.Commands.Bars.DeleteBar
 {
     public class DeleteBarCommandHandler : IRequestHandler<DeleteBarCommand>
     {
-        private readonly IDbContext _context;
+        private readonly ICrunchyPeanutButterDbContext _context;
 
-        public DeleteBarCommandHandler(IDbContext context)
+        public DeleteBarCommandHandler(ICrunchyPeanutButterDbContext context)
         {
             _context = context;
         }
@@ -17,6 +18,8 @@ namespace CrunchyPeanutButter.Core.Commands.Bars.DeleteBar
         public async Task<Unit> Handle(DeleteBarCommand request, CancellationToken cancellationToken)
         {
             var bar = await _context.Bars.FindAsync(request.Id);
+            if (bar == null)
+                throw new InvalidOperationException($"Couldn't find bar with ID {request.Id}.");
 
             _context.Bars.Remove(bar);
