+++ CrunchyPeanutButter.Core/Commands/Foos/DeleteFoo/DeleteFooCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using CrunchyPeanutButter.Core.Abstractions;
 using MediatR;
@@ -7,9 +8,9 @@ namespace CrunchyPeanutButter.Core.Commands.Foos.DeleteFoo
 {
     public class DeleteFooCommandHandler : IRequestHandler<DeleteFooCommand>
     {
-        private readonly IDbContext _context;
+        private readonly ICrunchyPeanutButterDbContext _context;
 
-        public DeleteFooCommandHandler(IDbContext context)
+        public DeleteFooCommandHandler(ICrunchyPeanutButterDbContext context)
         {
             _context = context;
         }
@@ -17,6 +18,8 @@ namespace CrunchyPeanutButter.Core.Commands.Foos.DeleteFoo
         public async Task<Unit> Handle(DeleteFooCommand request, CancellationToken cancellationToken)
         {
             var foo = await _context.Foos.FindAsync(request.Id);
+            if (foo == null)
+                throw new InvalidOperationException($"Couldn't find foo with ID {request.Id}.");
 
             _context.Foos.Remove(foo);
